add to_map function to vak.utils.labels
that maps a labelset to a dictionary. Has argument 'map_unlabeled'
that if True adds 'unlabeled' to the set and uses this "label" when
mapping the labelset from 0 to n labels.

Decide to do away with name 'silent_gap_label' and instead use
'unlabeled' which carries less baggage--you don't have to know
what a 'silent gap' is to understand it.

Also this now just forces the user to accept that the mapping will
be to consecutive integers from 0 to n; if for some reason users
don't want this they can specify their own mapping upon
instantiating a VocalizationDataset.