rewrite utils.labels.label_timebins
change first parameter name to 'labels_int' to emphasize that
labels should already have been mapped to integers, and add
pre-conditions that check whether this is the case then give a
specific error message, instead of the single line that was there
to cast each element of labels to int.

Also change 'silent_gap_label' parameter to 'unlabeled_label',
to agree with change in name in 'labels.to_map' function. Explain
this in docstring.

Rewrite/clean up rest of docstring as well.