rewrite VocalizationDataset defaults and validators for labelset and labelmap
labelset:
if not assigned, and there are annotations, then labelset defaults to the
actual unique set of labels found in the annotations.
The validator raises an error if there are labels in actual unique set of labels
that are not found in labelset. However it does not raise an error if the actual
unique set of labels is a ("true") subset of labelset (because maybe this is one
set of vocalizations in which the particular label did not occur but it would
occur in others and we want to denote that).

labelmap:
if not assigned, and labelset is None, then default to None.
if not assigned, and labelset is not None, figure out if there are unlabeled
timebins and if so make a labelmap that includes 'unlabeled' (probably this
will usually be the case); otherwise just make labelmap without 'unlabeled'.
The validator raises an error if the set of keys in labelmap (minus 'unlabeled')
does not equal labelset.