remove 'skip_files...' parameter from dataset.prep
so that this function, like dataset.audio and dataset.spect that it
calls, will maintain the same behavior as previously only when
labelset is defined. If labelset is None then no checking is done
about labels. This also makes it possible to make a dataset without
labels, .e.g for predicting labels from spectrograms