remove 'skip_files...' parameter from cli.prep
so that this function, like dataset.prep, will maintain the same behavior
as previously, but when only labelset is defined.
If labelset is None then no checking is done about labels.
This is also necessary to makes it possible to create a dataset without
labels, .e.g for predicting labels from spectrograms