fix bug in source_annot_map
Now make a copy of source_files list and pop items off that copy instead of
the original. Popping off the original was causing bugs in downstream
functions that use the original for other purposes.

But in this function we pop off items so we can check whether all source files
were matched to annotations--if there's no items in the list then we know we
matched them all. Making a copy is a cheap fix that lets use this approach.