change behavior of labelset_from_toml_value converter
Now always returns labelset as single-character strings
(e.g. {'a', 'b', '1', '3', '5'}). Handles strings like "abcde"
and a "range string" like "range: 1-20, 22" as before, but now
casts lists of ints to string, and parses each element of a
list of strings according to the rules for single strings--i.e.,
labelset=["range:1-3", "a-"] would become {'1', '2', '3', 'a', '-'}.