rewrite source_annot_map
so that it maps source files to annotations, not the other way around.

The function will no longer raise an error if it cannot map every
annotation in the input list `annot_list` to a source file in the
input list `source_files`.

Now it only raises an error if it cannot map every file in
`source_files` to an annotation (or some other pathological condition
occurs, e.g. multiple annotaiton files have audio_file attributes
that yield the same stem, which is what is used to match them with
the source files).