add io.util.is_valid_set_of_spect_files, fix #123 & #134
add util module to io, and add function is_valid_set_of_spect_files
to that module.
This is essentially a code block factored out of io.spect.to_dataframe.

The main difference is that instead of operating on `spect_annot_map`,
it takes a list of paths to spectrogram files--the annotations in the
spect_annot_map are not needed for this validation step.

Also the list of paths gets cast to Path objects first thing.
This fixes issue #123 where an error message used the .name attribute
on a `spect_path` that was actually a string.

In addition, parallelizes validation of spectrogram files using dask,
so it doesn't take forever on long datasets. Fixes #134.