rewrite _eval method of vak.model
so that it works with output of
EvalItemTransform, instead of working the
same way as the training data set.

At a high level: each "batch" is one item
from the dataset class, i.e. one spectrogram
transformed so it is a stack of windows.
We remove the batch dimension (of size 1),
get the predictions, flatten them, remove any
predictions made on padding, and then
use the result to compute any metrics like
accuracy.