add explicit close of progress bar to Model._train
When running learning curve, models appear to
continue training after the main loop.
But this is because the tqdm progress bar was
not explicitly closed before `break`ing out of
iterating through the training set
inside the `_train` method.

This adds an explicit `progress_bar.close()` just
before the `break` statement.