make brute_force seed each split with all labels
adds code before "inner loop" that iterates over
each label in labelset, and adds at least one
occurrence of a file that contains that label to
each split.

E.g. if the labelset is {'a','b','c'},
and the splits are ['train', 'val'],
the new code block makes sure that both
the 'train' and 'val' split contain
at least one file with 'a', 'b', and 'c'.

Testing on one canary with 30 labels (!), found
that this modified version of the function
worked 100% of the time while the old version
only worked 20% of the time