fix bugs in labels.lbl_tb2labels #239
the actual bug was because the input was a row, and wasn't cast
to a list. Just fixing the list didn't work though, because it
was then a list with one element, i.e. another list with many
elements. That's what gave the TypeError, the list comprehension
that converts integers in the list to labels was passing the
entire list into the dictionary to map from integers to labels,
but a list can't be a key (hence the "unhashable type" error).

- use 'row_or_1d' validator to ensure that input is 1-dimensional
  + chose to not make a 'row_or_column_or_1d' validator, in order
    to limit number of valid inputs, in case that might leave an
    opening for some other bug
- unindent labels.tolist() line
  + it was part of the code block that removes segments that had
    the 'unlabeled' label, but labels needs to be a list regardless