move/rename configs
decide it makes more sense to have the test configs
that will get copied into 'generated' inside the
'./tests/test_data/configs' folder. These files are super
small so it's not the end of the world to store them
in the repo.

and having them in "scripts" conflated config files with scripts

finally, rename so first word in config file name is 'test' not
'setup' since we're not really using them to "set up" anymore.
They can be run whenever we generate test data which is
decoupled from any "set up".