change behavior of labelset_from_toml_value, fix #277
make it possible to specify multi-character labels

behavior of ``vak.config.converts.labelset_from_toml_value``
function stays the same when the value for the ``labelset``
option in a toml file is a string, but when it is a list,
they are now processed like so:
any multi-character elementthat is not a "rangestr" now stays as is.
"rangestr" gets converted as before.
The function no longer requires the list to be
either all ints or all str.
That left open the possibility to fail silently,
because the converter function could just return a None
if the list was not all ints or all strs).
So now the function simply casts ints to string,
converts any "rangestr",
and leaves anything else (single/multi-character string) untouched.