make validation in `audio.to_spect` case-insensitive
`vak.io.audio.to_spect` validates that `audio_files`
supplied as argument all have the same extension,
and that they match the `audio_format` specified
when the function was called.

this caused an error when the format was `wav`
and the files all ended in `WAV`.
Fixed validation so that it is case-insensitive,
to match changes to `vak.files.from_dir`