add tests/test_data/configs/configs.json
with list of schematized config files,
that makes it possible for fixtures to "look up"
a specific config file by iterating
through the list and finding the one
that matches specified key-value pairs

other fixtures then piggy-back on this:
e.g. load toml from specific config,
get the csv_path from a certain section
of the toml from that specific config,
load that specific csv_path as a dataframe,
etc.