bug: fix how `lbl_tb2segments` computes on/offsets
currently the `lbl_tb2segments` function
approximates the onsets and offsets
in seconds by setting them to the time "between"
bins, i.e. half the width of the timebin;
this is the best estimate we can make,
given that we are assigning labels to each
timebin.

But when the input has an onset at index 0
or an offset at the last timebin, this produces
an estimate that is "outside" the valid times
in the vector of timebins.

Added a clean-up step that catches and fixes
these edge cases