BUG: fix how `spect.to_dataframe` handles annot_path
change `spect.to_dataframe` so that when there is no
annotation--e.g. when `prep`aring a dataset for
`predict`ion--it sets the `annot_path` to `np.nan`
instead of `None`

this prevents errors in unit testing where comparing
the dataframe returned by `core.prep` with the
dataframe loaded with `pd.read_csv` gives false positives
for "not equal"

and in general avoids needing to special-case
how this column is loaded. Just keep it as np.nan
when there's no annotation.
This is also faster than having a `pandas.Series` of
`None` which would be treated as `numpy.object` datatype.
Probably doesn't matter for most use cases
but might to start matter with giant datasets.