BUG: fix how `learncurve` handles `previous_run_path`, #316
When running `learncurve` with a `previous_run_path`,
we currently re-use the vectors that the `WindowSize`
dataset class uses to ensure the windows we grab our valid.
Those vectors are saved as .npy files on the previous run,
and right now the `train_dur_csv_paths_from_dir` function
just copies them to the new `results_dir`
so they get re-used.

But this causes a bug when we use a different window size.
Since which windows are valid depends on the window size,
of course.

The fix is to have `train_dur_csv_paths_from_dir` call
`WindowDataset.spect_vectors_from_df` instead of
using the saved `spect_vectors` from
the `previous_run_path`.

Since the `spect_vectors_from_df` method is
deterministic given the same `subset_df`,
this shouldn't change behavior when window size
is the same. But when window size changes,
we now generate the correct `spect_vectors`