BUG: fix Levenshtein distance implementation, fix #355
fixes a relatively minor bug in how Levenshtein distance
is computed

Originally we used a vectorized Numpy implementation,
the fifth algorithm under Python in this Wikibook:
https://en.wikibooks.org/wiki/Algorithm_Implementation/Strings/Levenshtein_distance#Python

but this implementation has a bug:
https://en.wikibooks.org/wiki/Talk:Algorithm_Implementation/Strings/Levenshtein_distance#Bug_in_vectorized_(5th)_version

I did not estimate how often this could cause a miscalculation.
But the previous implementation did pass all other tests I added --
only the example given in that discussion caused distance
to be off by 1.

I modified the implementation to match this version from `pylev`
https://github.com/toastdriven/pylev/blob/e79ef6814c51046fee635a2a2316f8a57fa1e697/pylev.py#L156

It now passes all tests including the case described
in that discussion.