BUG: Pass keys into spect.to_dataframe, fixes #412
In `vak.io.dataframe.from_files`,
pass keys from `spect_params`
into `spect.to_dataframe`
so that the correct keys are used
to access arrays in array files.
E.g., if user specified 'spect' as 'spect_key'
instead of the default 's',
in the 'spect_params',
be sure to use that instead of default.
Fixes #412.

Also, if `spect_params` is passed in to function
as an instance of `vak.config.spect_params.SpectParamsConfig`,
this gets converted to a `dict`
so that indexing by key works.