BUG: Add pre-condition for `labelset` in core.prep, fix #468
to avoid crash / silent error when purpose is
{'train', 'learncurve', 'eval'}.
Checks whether 'labelset' is None for any of these
"purposes" and if so raises ValueError,
with suggestion to add 'labelset' option to config.

For any "purpose" that results in a split,
not having this pre-condition results in an error
during the split, because you will end up with an
empty dataset from filtering out any vocalizations
with labels not in a labelset (that is None,
so that will be all files).

I don't think there would be an immediate error,
e.g. for eval where there's no split.
But I think it could end producing a silent error
because the evaluation dataset might have labels not
in the training set?

Later on will need to deal with cases
where we're not doing supervised learning,
but ignoring that for now