TST: Fix unit test in test_cli/test_learncurve.py
Fix `test_learning_curve_csv_path_none_raises`
to not cause spurious failures by
generating a new directory in
"tests/data_for_tests/generated/results/learncurve/data/model",
that other tests then see
because of how a fixture finds directories
('previous_run_path')
and try to use as test data.

Fix is to make a new root_results_dir
in the test set-up
and have the config point to it,
like all the other tests do.

A downside of this is
that it makes tests quite verbose.
Wonder if we could somehow
make the output_dir -> tmp_path
happen automatically for each test?