ENH: `prep` raises error when dataframe is empty, fix #116 [skip ci]
Have `vak.core.prep` raise ValueError
if `vak_df` returned by `dataframe.from_files` is empty.

I was unable to reliably find a way to cause
this error -- I should have put more detail in the
issue when I raised it.
So I am not adding a unit test for this.
But it still seems better to raise an error than to not,
since a user will be expecting the dataset
to not be "empty", and we don't want it to fail silently,
or cause an even more cryptic error downstream.