Add validation to `split.algorithms.brute_force`, fix #288
The function `brute_force` now validates that
the set of unique label classes in the list of
numpy arrays `labels` is equal to `labelset`,
and if not, raises an informative error message.

Adds two helper functions to algorithms/brute_force.py.
The first, `unique_set_from_labels`,
converts the list of numpy arrays `labels`
to a set of the unique label classes in the arrays.
The second, `validate_labels`, applies the
first to `labels` and then compares the result
to `labelset`, and raises an error if the two
do not match. The error message depends on
whether the set from labels is a proper subset
of `labelset` or vice versa, i.e., if labels is
missing some labels in `labelset` or if it has
extra labels not found in `labelset`.