CLN: Refactor redundant code in `vak.io.audio`
Removes repeated `if annot_list` code block
inside `audio.to_spect` function.

There were two places that check if there's a list of
annotations, and then call `map_annotated_to_annot`,
to map the annotation to `audio_files`.

The first in the `if audio_files` code block,
and the second in the `if audio_dir` code block
right under the first.

But after both code blocks we end up with a list of
`audio_files`.
So we can just wait until after and then check `if annot_list`
and (if there is an `annot_list`)
finally call `map_annotated_to_annot`.
This commit changes the function to do so.