CLN/MAINT: Change generated test data .tar filenames
in noxfile.py, so that we can distinguish
between test data for version 0.x and version 1.x.

Need to do this because CI is now failing
when it tries to download a non-existent .tar.gz file,
since I uploaded a new one for testing 1.0.

In practice there might not be a big enough difference
yet for it to affect tests
(I got a PR for version 0.x to pass tests locally
even though I was using data generated for 1.0)
but at some point they will diverge.

I think for now crudely dividing up by major version
is good enough.