BUG: Fix how `model` decorator sets subclass' module
Multiprocessing fails when it tries to pickle
instances of a subclass made with the
`vak.models.model` decorator, i.e. TweetyNet
and TeenyTweetyNet.

This is because we were giving all attributes
from the family we were subclassing to the subclass,
including `__module__`.
So the subclass would appear to have the `__module__`
of the superclass, and when pickle tried to
do attribute lookup on that module, it would fail.

The fix is to directly set `subclass.__module`
to `definition.__module__` after making the subclass
with type. After doing so, multiprocessing works.