Rename annotation.from_df parameter `dataset_path` to `annot_root`
Make it optional but have all internal functions use it.

Needed to do this because unit test calls this function to test
output of `io.dataframe.from_files`.

Passing in any value "worked" because the paths were absolute,
but really we should be able to get the annotations
from the dataframe at any point.

Minor detail but I don't want this to be confusing later.