Rewrite `vak.prep.learncurve.make_learncurve_splits_from_dataset_df`
We now make one directory in `dataset_path` for every split,
and save arrays in those.
This means `WindowDataset.from_dataset_path` can work the same way
for these splits/folds used in a learning curve and the overall
splits of the dataset, instead of needing to add logic that special
cases learning curve.
It works because the learncurve split name becomes a sub-directory of
`dataset_path`, just like `train`/`test`/`val` splits.
We no longer make a dataframe / csv for each split and then make
a json file where the keys are training set durations + replicates
and the values are the csv paths.
Instead we make a single dataframe containing all the split
dataframes, and we save that over the original `dataset_df` passed in,
located at `csv_path`.