Fix summation of skip connections in TCNBlock
The summation was originally written to include `x`.
The intent was to include the *input*
but actually the way it was written,
we were including the output of the last
Residual block.
We should not do this;
instead we *only* add the output of each
skip connection--that's what DAS does,
as confirmed by inspecting the model summary
in a Jupyter notebook.
This is similar to what WaveNet does
(without the additional 1x1 convolutions)
as well as SELD-TCN.