Make `prep_frame_classification_dataset` patch "source_file" column in dataframe
Add logic that forces the `source_file` column to point at npy files
that contain spectrograms.

I think this springs from a misunderstanding of what the source_file column is,
probably needs to be better documented.
Patching this shouldn't change anything since the main dataset class doesn't
look at this column when it is loading data.
Unless the dataset class was also modified.